updates to latest softfloat

+++ hwacha/insns_ut/vcmpfeq_h.h
@@ -1,3 +1,3 @@
 require_fp;
-WRITE_PRD(f32_eq(HFRS1, HFRS2));
+WRITE_PRD(f32_eq(f32(HFRS1), f32(HFRS2)));
 set_fp_exceptions;
