+++ hwacha/insns_ut/vcmpfeq_s.h
@@ -1,3 +1,3 @@
 require_fp;
-WRITE_PRD(f32_eq(FRS1, FRS2));
+WRITE_PRD(f32_eq(f32(FRS1), f32(FRS2)));
 set_fp_exceptions;
