+++ hwacha/insns_ut/vcmpfle_d.h
@@ -1,3 +1,3 @@
 require_fp;
-WRITE_PRD(f64_le(FRS1, FRS2));
+WRITE_PRD(f64_le(f64(FRS1), f64(FRS2)));
 set_fp_exceptions;
