+++ hwacha/insns_ut/vcmpflt_s.h
@@ -1,3 +1,3 @@
 require_fp;
-WRITE_PRD(f32_lt(FRS1, FRS2));
+WRITE_PRD(f32_lt(f32(FRS1), f32(FRS2)));
 set_fp_exceptions;
