+++ hwacha/insns_ut/vfclass_d.h
@@ -1,2 +1,2 @@
 require_fp;
-WRITE_RD(f64_classify(FRS1));
+WRITE_RD(f64_classify(f64(FRS1)));
