+++ hwacha/insns_ut/vfclass_h.h
@@ -1,2 +1,2 @@
 require_fp;
-WRITE_RD(f32_classify(HFRS1));
+WRITE_RD(f32_classify(f32(HFRS1)));
