+++ hwacha/insns_ut/vfclass_s.h
@@ -1,2 +1,2 @@
 require_fp;
-WRITE_RD(f32_classify(FRS1));
+WRITE_RD(f32_classify(f32(FRS1)));
