+++ hwacha/insns_ut/vfcvt_l_s.h
@@ -1,5 +1,5 @@
 require_rv64;
 require_fp;
 softfloat_roundingMode = VRM;
-WRITE_RD(f32_to_i64(FRS1, VRM, true));
+WRITE_RD(f32_to_i64(f32(FRS1), VRM, true));
 set_fp_exceptions;
