+++ hwacha/insns_ut/vfcvt_lu_h.h
@@ -1,5 +1,5 @@
 require_rv64;
 require_fp;
 softfloat_roundingMode = VRM;
-WRITE_RD(f32_to_ui64(HFRS1, VRM, true));
+WRITE_RD(f32_to_ui64(f32(HFRS1), VRM, true));
 set_fp_exceptions;
