+++ hwacha/insns_ut/vfcvt_w_h.h
@@ -1,4 +1,4 @@
 require_fp;
 softfloat_roundingMode = VRM;
-WRITE_RD(sext32(f32_to_i32(HFRS1, VRM, true)));
+WRITE_RD(sext32(f32_to_i32(f32(HFRS1), VRM, true)));
 set_fp_exceptions;
