+++ hwacha/insns_ut/vfcvt_wu_s.h
@@ -1,4 +1,4 @@
 require_fp;
 softfloat_roundingMode = VRM;
-WRITE_RD(sext32(f32_to_ui32(FRS1, VRM, true)));
+WRITE_RD(sext32(f32_to_ui32(f32(FRS1), VRM, true)));
 set_fp_exceptions;
