+++ hwacha/insns_ut/vfmax_h.h
@@ -1,4 +1,4 @@
 require_fp;
-WRITE_HFRD(isNaNF32UI(HFRS2) || f32_le_quiet(HFRS2,HFRS1) /* && FRS1 not NaN */
+WRITE_HFRD(isNaNF32UI(HFRS2) || f32_le_quiet(f32(HFRS2),f32(HFRS1)) /* && FRS1 not NaN */
       ? HFRS1 : HFRS2);
 set_fp_exceptions;
