+++ hwacha/insns_ut/vfmin_d.h
@@ -1,4 +1,4 @@
 require_fp;
-WRITE_FRD(isNaNF64UI(FRS2) || f64_lt_quiet(FRS1,FRS2) /* && FRS1 not NaN */
+WRITE_FRD(isNaNF64UI(FRS2) || f64_lt_quiet(f64(FRS1),f64(FRS2)) /* && FRS1 not NaN */
       ? FRS1 : FRS2);
 set_fp_exceptions;
